fix: cannot jump to query from dashboard
